Added UTLFedoraObject class
This class is used as an interface to the repository. It used to exist as a library imported with Drupal's libraries module.
This modification includes UTLFedoraObject as a part of SC Manifest module for redistribution.
In addition, the hardcoded UTL URLs are replaced by hardcoded localhost. These URLs should ideally be Drupal options.

Also, fixed another reference sign in function call.